separate composables for move, resize & new entry
